+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultEventAnalyticsController.kt
@@ -58,6 +58,7 @@ class DefaultEventAnalyticsController(
             eventAttributes[EventAttributeName.batteryLevel] = it
         }
         eventAttributes[EventAttributeName.batteryState] = appStateMonitor.getBatteryState().description
+        eventAttributes[EventAttributeName.lowPowerModeEnabled] = appStateMonitor.isBatterySaverOn().toString()
 
         eventReporter?.report(SDKEvent(name, eventAttributes))
 
@@ -73,7 +74,8 @@ class DefaultEventAnalyticsController(
         val eventAttributes = mutableMapOf(
             EventAttributeName.timestampMs to currentTimeMs,
             EventAttributeName.appState to appStateMonitor.appState.description,
-            EventAttributeName.batteryState to appStateMonitor.getBatteryState().description
+            EventAttributeName.batteryState to appStateMonitor.getBatteryState().description,
+            EventAttributeName.lowPowerModeEnabled to appStateMonitor.isBatterySaverOn().toString()
         ) as EventAttributes
 
         appStateMonitor.getBatteryLevel()?.let {