+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/video/DefaultVideoClientController.kt
@@ -76,11 +76,9 @@ class DefaultVideoClientController(
             DefaultCameraCaptureSource(
                 context,
                 logger,
-                surfaceTextureCaptureSourceFactory
-            ).apply {
-                eventAnalyticsController =
-                    this@DefaultVideoClientController.eventAnalyticsController
-            }
+                surfaceTextureCaptureSourceFactory,
+                eventAnalyticsController
+            )
     }
 
     private var videoClient: VideoClient? = null