+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/device/DefaultDeviceControllerTest.kt
@@ -13,7 +13,6 @@ import android.media.AudioRecordingConfiguration
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsController
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
-import com.amazonaws.services.chime.sdk.meetings.analytics.MeetingHistoryEventName
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.AudioClientController
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.AudioClientState
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.DefaultAudioClientController
@@ -143,22 +142,6 @@ class DefaultDeviceControllerTest {
     fun tearDown() {
         Dispatchers.resetMain()
         testDispatcher.cleanupTestCoroutines()
-    }
-
-    @Test
-    fun `deviceController should call addMeetingHistoryEvent when audio device is selected`() {
-        setupForNewAPILevel()
-        every { audioClientController.setRoute(any()) } returns true
-        mockkStatic(DefaultAudioClientController::class)
-        DefaultAudioClientController.audioClientState = AudioClientState.STARTED
-        deviceController.chooseAudioDevice(MediaDevice(
-            "speaker",
-            MediaDeviceType.AUDIO_BUILTIN_SPEAKER
-        ))
-
-        verify(exactly = 1) {
-            eventAnalyticsController.pushHistory(MeetingHistoryEventName.audioInputSelected)
-        }
     }
 
     @Test
@@ -375,6 +358,42 @@ class DefaultDeviceControllerTest {
         )
 
         verify { audioClientController.setRoute(AudioClient.SPK_STREAM_ROUTE_RECEIVER) }
+    }
+
+    @Test
+    fun `chooseAudioDevice should call publishEvent when setRoute`() {
+        setupForOldAPILevel()
+        every { audioClientController.setRoute(any()) } returns true
+        mockkStatic(DefaultAudioClientController::class)
+        DefaultAudioClientController.audioClientState = AudioClientState.STARTED
+        deviceController.chooseAudioDevice(
+            MediaDevice(
+                "usb headset",
+                MediaDeviceType.AUDIO_USB_HEADSET
+            )
+        )
+
+        verify { eventAnalyticsController.publishEvent(EventName.audioInputSelected, mutableMapOf(
+            EventAttributeName.audioDeviceType to MediaDeviceType.AUDIO_USB_HEADSET.toString()
+        ), false) }
+    }
+
+    @Test
+    fun `deviceController should call publishEvent when audio device is selected`() {
+        setupForNewAPILevel()
+        every { audioClientController.setRoute(any()) } returns true
+        mockkStatic(DefaultAudioClientController::class)
+        DefaultAudioClientController.audioClientState = AudioClientState.STARTED
+        deviceController.chooseAudioDevice(MediaDevice(
+            "speaker",
+            MediaDeviceType.AUDIO_BUILTIN_SPEAKER
+        ))
+
+        verify(exactly = 1) {
+            eventAnalyticsController.publishEvent(EventName.audioInputSelected, mutableMapOf(
+                EventAttributeName.audioDeviceType to MediaDeviceType.AUDIO_BUILTIN_SPEAKER.toString()
+            ), false)
+        }
     }
 
     @Test