+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/internal/video/DefaultContentShareVideoClientObserverTest.kt
@@ -6,12 +6,16 @@
 package com.amazonaws.services.chime.sdk.meetings.internal.video
 
 import android.content.Context
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsController
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare.ContentShareObserver
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare.ContentShareStatus
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare.ContentShareStatusCode
 import com.amazonaws.services.chime.sdk.meetings.internal.contentshare.DefaultContentShareVideoClientObserver
 import com.amazonaws.services.chime.sdk.meetings.internal.metric.ClientMetricsCollector
 import com.amazonaws.services.chime.sdk.meetings.session.URLRewriter
+import com.amazonaws.services.chime.sdk.meetings.utils.VideoClientFailedError
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import com.xodee.client.video.VideoClient
 import io.mockk.MockKAnnotations
@@ -48,6 +52,9 @@ class DefaultContentShareVideoClientObserverTest {
     @MockK
     private lateinit var mockURLRewriter: URLRewriter
 
+    @MockK
+    private lateinit var mockEventAnalyticsController: EventAnalyticsController
+
     @InjectMockKs
     private lateinit var testContentShareVideoClientObserver: DefaultContentShareVideoClientObserver
 
@@ -59,7 +66,7 @@ class DefaultContentShareVideoClientObserverTest {
 
     private val testDispatcher = TestCoroutineDispatcher()
 
-    private val videoClientStatus = 0
+    private val videoClientStatus = 6
 
     @Before
     fun setUp() {
@@ -79,6 +86,16 @@ class DefaultContentShareVideoClientObserverTest {
     }
 
     @Test
+    fun `didConnect should call publishEvent with ContentShareStarted`() {
+        testContentShareVideoClientObserver.didStop(mockVideoClient)
+
+        verify(exactly = 1) {
+            mockContentShareObserver.onContentShareStarted()
+            mockEventAnalyticsController.publishEvent(EventName.contentShareStarted)
+        }
+    }
+
+    @Test
     fun `didFail should notify observer of onContentShareStopped event with VideoServiceFailed status`() {
         testContentShareVideoClientObserver.didFail(
             mockVideoClient,
@@ -92,6 +109,9 @@ class DefaultContentShareVideoClientObserverTest {
                     ContentShareStatusCode.VideoServiceFailed
                 )
             )
+            mockEventAnalyticsController.publishEvent(EventName.contentShareFailed, mutableMapOf(
+                EventAttributeName.contentShareErrorMessage to VideoClientFailedError.fromVideoClientStatus(videoClientStatus)
+            ))
         }
     }
 
@@ -124,6 +144,9 @@ class DefaultContentShareVideoClientObserverTest {
         testContentShareVideoClientObserver.didStop(mockVideoClient)
 
         verify { mockClientMetricsCollector.processContentShareVideoClientMetrics(emptyMap()) }
+        verify(exactly = 1) {
+            mockEventAnalyticsController.publishEvent(EventName.contentShareStopped)
+        }
     }
 
     @Test